# Small microptimization to year_index.