@tachyonwill 2021-11-30
Ah, that's unfortunate. Would making this C++14 or higher via macro be acceptable? Something like(to use an example from abseil):
``
#if defined(__cplusplus) && __cplusplus >= 201402L
...
```