@devbww 2021-11-30
Allow me to attempt to clear up a little confusion.  `year_index()` is already non-constexpr unless `__cpp_constexpr >= 201304`.