@devbww 2021-11-30
And that said, it may be possible to expand the `year_index()` result range from `[0 .. 399]` to `[-399 .. 399]`, and avoid the need for the `((... + 400) % 400)` altogether, by making the two places it is called a little more conditional.