@tachyonwill 2021-11-30
It sounds like an additional macro is not needed then. Changing days_per_4years to handle negatives could be done but I think would be ugly. I had another idea. What if we just compute year_index once in the n_day loop section and just maintain it through addition and a conditional subtraction? In the application where we noticed this code coming up in profiling, it is really the stepping through 4 years at a time part of the code that was problematic, so caching there might be helpful. I am will rerun the benchmarks with this change.