@devbww 2021-11-30
> it may be possible to expand the `year_index()` result range from `[0 .. 399]` to `[-399 .. 399]`

For example, ...
```
CONSTEXPR_F int new_year_index(year_t y, month_t m) noexcept {
  return static_cast<int>((y + (m > 2)) % 400);
}
CONSTEXPR_F int new_days_per_century(year_t y, month_t m) noexcept {
  const int yi = new_year_index(y, m);
  return 36524 + ((yi > -100 && yi <= 0) || yi > 300);
}
CONSTEXPR_F int new_days_per_4years(year_t y, month_t m) noexcept {
  const int yi = new_year_index(y, m);
  return 1460 + ((yi > -100 && yi <= 0) || yi > 300 || (yi + 399) % 100 < 96);
}
```