@tachyonwill 2021-11-30
This does look a little better, particularly BM_Time_ToCivilUTC_CCTZ.