@tachyonwill 2021-11-30
I didn't see that latest comment. I agree that is how you would expand those conditionals to handle negative modulus but I think think it reduces code legibility? The logic is already pretty cryptic. It's why I didn't go with that approach. I've updated the CL to the approach where we don't recompute year_index from scratch on each iteration.