@devbww 2021-12-01
@tachyonwill ... Before we forget about this, do you think it might be worth avoiding the `year_index()` modulus in the common (contemporary) cases with something like ...
```
CONSTEXPR_F int new_year_index(year_t y, month_t m) noexcept {
  const year_t ey = y + (m > 2);
  if (ey < 2400) {
    if (ey >= 2000) return static_cast<int>(ey - 2000);
    if (ey >= 1600) return static_cast<int>(ey - 1600);
  }
  const int yi = static_cast<int>(ey % 400);
  return yi < 0 ? yi + 400 : yi;
}
```
I guess any benefit is probably unobservable now that `year_index()` is only called once per civil-time normalization, but you might have a better intuition.