@tachyonwill 2021-12-02
I'm not sure, but I would side on it not being worth it. I did some not very rigorous tests that suggested that if almost all of your data is before 2000 or almost all after 2000 this could be beneficial in the abstract. Once your data starts splitting that boundary, performance actually drops slightly. I don't think IRL it will matter much either way for the reason you mentioned.